old compiler and os fix

+++ apps/ech.c
@@ -257,7 +257,7 @@ int ech_main(int argc, char **argv)
                 goto end;
             }
             BIO_printf(bio_err, "ECH entry: %d public_name: %s age: %lld%s%s\n",
-                oi_ind, pn, (long long)secs,
+                oi_ind, pn, (int64_t)secs,
                 has_priv ? " (has private key)" : "",
                 for_retry ? " (will be sent in retry-configs)" : "");
             BIO_printf(bio_err, "\t%s\n", ec);
